// src/features/CreateNewPage/model/consts/consts.ts
export enum ValidateNewArticleError {
    NO_DATA = 'NO_DATA',
    SERVER_ERROR_SAVE = 'SERVER_ERROR_SAVE',
    SERVER_ERROR_GET = 'SERVER_ERROR_GET',
    INCORRECT_TITLE = 'INCORRECT_TITLE',
    INCORRECT_SUBTITLE = 'INCORRECT_SUBTITLE',
    INCORRECT_IMG_URL_EMPTY = 'INCORRECT_IMG_URL_EMPTY',
    INCORRECT_IMG_URL_INVALID = 'INCORRECT_IMG_URL_INVALID',
    INCORRECT_NEW_BLOCKS_LENGTH = 'INCORRECT_NEW_BLOCKS_LENGTH',
    INCORRECT_NEW_TEXT_BLOCK = 'INCORRECT_NEW_TEXT_BLOCK',
    INCORRECT_NEW_CODE_BLOCK = 'INCORRECT_NEW_CODE_BLOCK',
    INCORRECT_NEW_IMAGE_BLOCK_TITLE = 'INCORRECT_NEW_IMAGE_BLOCK_TITLE',
    INCORRECT_IMG_URL_SRC_EMPTY = 'INCORRECT_IMG_URL_SRC_EMPTY',
    INCORRECT_IMG_URL_SRC_INVALID = 'INCORRECT_IMG_URL_SRC_INVALID',
}